osal Integration candidate: 2021-02-02

+++ src/tests/queue-test/queue-test.c
@@ -35,7 +35,7 @@ void QueueTimeoutCheck(void);
 #define MSGQ_DEPTH 50
 #define MSGQ_SIZE  sizeof(uint32)
 #define MSGQ_TOTAL 10
-#define MSGQ_BURST  3
+#define MSGQ_BURST 3
 
 /* Task 1 */
 #define TASK_1_STACK_SIZE 1024
@@ -67,8 +67,8 @@ void task_1(void)
 {
     int32  status;
     size_t data_size;
-    uint32   data_received ;
-    uint32   expected = 0;
+    uint32 data_received;
+    uint32 expected = 0;
 
     OS_printf("Starting task 1\n");
 
@@ -87,7 +87,7 @@ void task_1(void)
         {
             ++task_1_messages;
             UtAssert_True(data_received == expected, "TASK 1: data_received (%u) == expected (%u)",
-                          data_received, expected);
+                          (unsigned int)data_received, (unsigned int)expected);
 
             expected++;
         }
@@ -198,12 +198,12 @@ void QueueMessageSetup(void)
 {
     int32  status;
     uint32 accuracy;
-    int        i;
-    uint32 Data = 0;
+    int    i;
+    uint32 Data     = 0;
     task_1_failures = 0;
     task_1_messages = 0;
     task_1_timeouts = 0;
-  
+
     status = OS_QueueCreate(&msgq_id, "MsgQ", OSAL_BLOCKCOUNT_C(MSGQ_DEPTH), OSAL_SIZE_C(MSGQ_SIZE), 0);
     UtAssert_True(status == OS_SUCCESS, "MsgQ create Id=%lx Rc=%d", OS_ObjectIdToInteger(msgq_id), (int)status);
 
@@ -227,16 +227,16 @@ void QueueMessageSetup(void)
     status = OS_TimerSet(timer_id, timer_start, timer_interval);
     UtAssert_True(status == OS_SUCCESS, "Timer 1 set Rc=%d", (int)status);
 
-    /* 
-    * Put 10 messages onto the que with some time inbetween the later messages   
-    * to make sure the que handles both storing and waiting for messages 
-    */
+    /*
+     * Put 10 messages onto the que with some time inbetween the later messages
+     * to make sure the que handles both storing and waiting for messages
+     */
     for (i = 0; i < MSGQ_TOTAL; i++)
     {
-        if(i > MSGQ_BURST)
+        if (i > MSGQ_BURST)
             OS_TaskDelay(400);
 
-        Data = i;
+        Data   = i;
         status = OS_QueuePut(msgq_id, (void *)&Data, sizeof(Data), 0);
         UtAssert_True(status == OS_SUCCESS, "OS Queue Put Rc=%d", (int)status);
     }
